[intermcu] subsystems -> subsystems
